requirements.txt
appdirs==1.4.4
astunparse==1.6.3
autograd==1.8.0
autoray==0.8.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
diastatic-malt==2.15.2
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pennylane==0.43.0
pennylane_lightning==0.43.0
pillow==12.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rustworkx==0.17.1
scikit-learn==1.7.2
scipy==1.16.2
scipy-openblas32==0.3.30.0.2
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.9.0
typer==0.19.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
